New implementation of element rotation
